authorization implemented

+++ src/api/songs/handler.js
@@ -1,5 +1,3 @@
-/* eslint-disable no-console */
-/* eslint-disable no-underscore-dangle */
 const ClientError = require('../../exceptions/ClientError');
 
 class SongsHandler {
@@ -20,9 +18,10 @@ class SongsHandler {
       const {
         title = 'untitled', year, performer, genre, duration,
       } = request.payload;
+      const { id: credentialId } = request.auth.credentials;
 
       const songId = await this._service.addSong({
-        title, year, performer, genre, duration,
+        title, year, performer, genre, duration, owner: credentialId,
       });
 
       const response = h.response({
@@ -54,8 +53,9 @@ class SongsHandler {
     }
   }
 
-  async getSongsHandler() {
-    const songs = await this._service.getSongs();
+  async getSongsHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const songs = await this._service.getSongs(credentialId);
     return {
       status: 'success',
       data: {
@@ -67,6 +67,9 @@ class SongsHandler {
   async getSongByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifySongOwner(id, credentialId);
       const song = await this._service.getSongById(id);
       return {
         status: 'success',
@@ -100,6 +103,9 @@ class SongsHandler {
       this._validator.validateSongPayload(request.payload);
 
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifySongOwner(id, credentialId);
 
       await this._service.editSongById(id, request.payload);
 
@@ -130,6 +136,9 @@ class SongsHandler {
   async deleteSongByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifySongOwner(id, credentialId);
       await this._service.deleteSongById(id);
       return {
         status: 'success',
